Treat assembler options in braces as optional limitation on the used instruction set rather than enforced target qualifiers.
It does not make sense to put such limitations in the code -- it is rather a limitation that should be directed by the compiler (or user of the compiler). Providing different targets in the source code only makes sense when different execution modes are possible in one architecture. This is still possible on an optional basis.

git-svn-id: https://svn.inf.ethz.ch/svn/lecturers/a2/trunk@5210 8c9fc860-2736-0410-a75d-ab315db34111