Implemented release lock upon trap in exclusive region (and raise a trap again).
A trap within (particularly within an exclusive section) is considered a programming error and leaving the object in an inconsistent state is unacceptable.
However, leaving the object in an inconsistent state AND locked is usually worse. We choose the smaller of the two evils in order to keep the system (in particular the window manager) in a potentially recoverable state. This also leaves the possibility to patch bugs in a running system.

git-svn-id: https://svn.inf.ethz.ch/svn/lecturers/a2/trunk@7639 8c9fc860-2736-0410-a75d-ab315db34111