Added basic support for conditional compilation
The compiler accepts a new optional called --define which contains a comma-separated list of identifiers which can be used to conditionally compile blocks of arbitrary code according to the following syntax:

    Block = '#' 'if' Expression 'then' Block { '#' 'elsif' Expression 'then' Block } ['#' 'else' Block] '#' 'end' | any other token until next new line.
    Expression = Term {'or' Term}.
    Term = Factor {'&' Factor}.
    Factor = Identifier | '~' Factor | '(' Expression ')'.

Identifiers in conditional if and elsif statements yield a Boolean value indicating whether they have been defined when invoking the compiler, but they are not available in normal Oberon code. Conditional statements begin with a # sign on separate lines and are designed to combine code blocks for different platforms and architectures; They are not meant to provide a generic macro replacement mechanism as preprocessors do.

git-svn-id: https://svn.inf.ethz.ch/svn/lecturers/a2/trunk@8167 8c9fc860-2736-0410-a75d-ab315db34111