support for return by reference
rationale: currently, user defined index operators have to be implemented in a writing and a reading form. Moreover, the compiler has to emit write back when the result of an index operator shall be passed by reference to a function. 

Instead, return by reference allows to write to a reference

OPERATOR "[]"(VAR s: String; index: SIZE): VAR CHAR;
BEGIN
	RETURN s.array[index]; (* return by reference *)
...
END "[]";

str[10] := 'a';

CallProcedureWithVarPar(str[8]);

It is important that the return by reference parameter cannot refer to objects that disappear from the stack. This will be ensured by allowing only references to paramters passed in and by disallowing passing references further out. This check is not yet implemented.

Return by reference is considered in an experimental state and is subject to change.

A distinction between CONST and VAR parameters in the signature has been added with the previous commit for handling of const objects.

git-svn-id: https://svn.inf.ethz.ch/svn/lecturers/a2/trunk@9726 8c9fc860-2736-0410-a75d-ab315db34111